fix: PlatformImage addFrame - fillFrame
- addFrame and fillFrame can now run together in any order.
- test does not check if result is equal to expected image.
- Missing pixel test, count size is not enough.
- Fix test image size for each platform so tests are  passing.

Explanation:
[ChatGPT](https://chatgpt.com/share/67bcdbbd-ec90-8008-b78b-4e4ae15bdbda)
Ensure Bitmap Representation: When drawing the border with addFrame(), ensure that an NSBitmapImageRep is available so that subsequent operations like fillFrame() can work with the image.
Avoid Direct Bitmap Representation Access in addFrame: Instead of trying to find the bitmapRep in the addFrame() method, you can work directly with NSImage's lockFocus() and unlockFocus() operations, and then use those images directly.